requirements.txt
Flask==2.0.1
google-cloud-aiplatform==1.23.0
google-cloud-storage==1.42.0
py-ogp-parser==0.2.1
